@@ -2,7 +2,7 @@ import { DebugElement } from '@angular/core';
 import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { By } from '@angular/platform-browser';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { SharedModule } from './../../shared/shared.module';
 import { MimeViewerIntl } from './../../core/intl/viewer-intl';
