@@ -1,6 +1,5 @@
 import { Injectable, ElementRef } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { Observable, ReplaySubject } from 'rxjs';
 
 import { MimeDomHelper } from '../mime-dom-helper';
 import { Dimensions } from '../models/dimensions';
