@@ -75,6 +75,9 @@ return (
             '@angular/common/testing': 'ng.common.testing',
             '@angular/common/http/testing': 'ng.common.http.testing',
             '@angular/flex-layout': 'ng.flex-layout',
+            '@angular/flex-layout/core': 'ng.flexLayout.core',
+            '@angular/flex-layout/extended': 'ng.flexLayout.extended',
+            '@angular/flex-layout/core': 'ng.flexLayout.flex',
 
             // Some packages are not really needed for the UMD bundles, but for the missingRollupGlobals rule.
             '@angular/material': 'ng.material',
@@ -82,48 +85,8 @@ return (
 
             // Include secondary entry-points of the cdk and material packages
 
-            'rxjs/BehaviorSubject': 'Rx',
-            'rxjs/ReplaySubject': 'Rx',
-            'rxjs/Observable': 'Rx',
-            'rxjs/Subject': 'Rx',
-            'rxjs/Subscription': 'Rx',
-            'rxjs/Observer': 'Rx',
-            'rxjs/Subscriber': 'Rx',
-            'rxjs/Scheduler': 'Rx',
-
-            'rxjs/observable/combineLatest': 'Rx.Observable',
-            'rxjs/observable/forkJoin': 'Rx.Observable',
-            'rxjs/observable/fromEvent': 'Rx.Observable',
-            'rxjs/observable/merge': 'Rx.Observable',
-            'rxjs/observable/of': 'Rx.Observable',
-            'rxjs/observable/throw': 'Rx.Observable',
-            'rxjs/observable/defer': 'Rx.Observable',
-            'rxjs/observable/fromEventPattern': 'Rx.Observable',
-            'rxjs/observable/empty': 'Rx.Observable',
-
-            'rxjs/operators': 'Rx.Observable',
-            'rxjs/operators/debounceTime': 'Rx.Observable',
-            'rxjs/operators/takeUntil': 'Rx.Observable',
-            'rxjs/operators/first': 'Rx.Observable',
-            'rxjs/operators/filter': 'Rx.Observable',
-            'rxjs/operators/map': 'Rx.Observable',
-            'rxjs/operators/tap': 'Rx.Observable',
-            'rxjs/operators/startWith': 'Rx.Observable',
-            'rxjs/operators/auditTime': 'Rx.Observable',
-            'rxjs/operators/switchMap': 'Rx.Observable',
-            'rxjs/operators/finalize': 'Rx.Observable',
-            'rxjs/operators/catchError': 'Rx.Observable',
-            'rxjs/operators/share': 'Rx.Observable',
-            'rxjs/operators/delay': 'Rx.Observable',
-            'rxjs/operators/combineLatest': 'Rx.Observable',
-            'rxjs/operators/throttle': 'Rx.Observable',
-            'rxjs/operators/interval': 'Rx.Observable',
-            'rxjs/operators/val': 'Rx.Observable',
-            'rxjs/operators/take': 'Rx.Observable',
-            'rxjs/operators/sample': 'Rx.Observable',
-            'rxjs/operators/distinctUntilChanged': 'Rx.Observable',
-            'rxjs/operators/finally': 'Rx.Observable',
-            'rxjs/observable/interval': 'Rx.Observable'
+            rxjs: 'Rx',
+            'rxjs/operators': 'Rx.Observable'
           },
           sourcemap: true
         },
@@ -141,20 +104,11 @@ return (
           '@angular/forms',
           '@angular/material',
           '@angular/flex-layout',
+          '@angular/flex-layout/core',
+          '@angular/flex-layout/extended',
+          '@angular/flex-layout/flex',
           'openseadragon',
-          'rxjs/Observable',
-          'rxjs/BehaviorSubject',
-          'rxjs/Subject',
-          'rxjs/ReplaySubject',
-          'rxjs/Subscription',
-          'rxjs/observable/interval',
-          'rxjs/operators/throttle',
-          'rxjs/operators/filter',
-          'rxjs/operators/distinctUntilChanged',
-          'rxjs/operators/finalize',
-          'rxjs/operators/take',
-          'rxjs/operators/takeUntil',
-          'rxjs/operators/sample',
+          'rxjs',
           'rxjs/operators'
         ],
         plugins: [sourcemaps()],
