@@ -1,14 +1,12 @@
-import { defineSupportCode } from 'cucumber';
-import { expect } from 'chai';
+const { Then } = require('cucumber');
+const { expect } = require('chai');
 
 import { ViewerPage } from './../pages/viewer.po';
 import { MetadataPage } from './../pages/metadata.po';
 
-defineSupportCode(function({ Given, When, Then }) {
-  const viewer = new ViewerPage();
-  const metadata = new MetadataPage();
+const viewer = new ViewerPage();
+const metadata = new MetadataPage();
 
-  Then('the logo associated with the resource are displayed to the user', async () => {
-    expect(await metadata.isLogoDisplayed()).to.equal(true);
-  });
+Then('the logo associated with the resource are displayed to the user', async () => {
+  expect(await metadata.isLogoDisplayed()).to.equal(true);
 });
