@@ -6,11 +6,11 @@ const http = require('http');
 const app = express();
 
 // Point static path to dist
-app.use(express.static(path.join(__dirname, '../dist')));
+app.use(express.static(path.join(__dirname, '../dist/integration')));
 
 // Catch all other routes and return the index file
 app.get('*', (req, res) => {
-  res.sendFile(path.join(__dirname, '../dist/index.html'));
+  res.sendFile(path.join(__dirname, '../dist/integration/index.html'));
 });
 
 /**
